Move join game modal to modal folder
